Fix rare 64 bit windows crash in asm

+++ src/core/exprfilter.cpp
@@ -96,7 +96,7 @@ typedef struct {
 } ExprData;
 
 #ifdef VS_TARGET_CPU_X86
-extern "C" void vs_evaluate_expr_sse2(const void *exprs, const uint8_t **rwptrs, const intptr_t *ptroffsets, int numiterations, void *stack);
+extern "C" void vs_evaluate_expr_sse2(const void *exprs, const uint8_t **rwptrs, const intptr_t *ptroffsets, intptr_t numiterations, void *stack);
 #endif
 
 static void VS_CC exprInit(VSMap *in, VSMap *out, void **instanceData, VSNode *node, VSCore *core, const VSAPI *vsapi) {
